requirements.txt
pyyaml>=6.0
sounddevice>=0.4.6
numpy>=1.24
mutagen>=1.47
websockets>=12.0